@@ -4,12 +4,12 @@ import 'bloc.dart';
 class Provider extends InheritedWidget {
   final bloc = Bloc();
 
-  Provider({Key key, Widget child}) : super(key: key, child: child);
+  Provider({Key? key, required Widget child}) : super(key: key, child: child);
 
   @override
   bool updateShouldNotify(_) => true;
 
   static Bloc of(BuildContext context) {
-    return context.dependOnInheritedWidgetOfExactType<Provider>().bloc;
+    return context.dependOnInheritedWidgetOfExactType<Provider>()!.bloc;
   }
 }
