@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:learn/src/animation/app.dart';
+import 'package:learn/src/news/app.dart';
 
 void main() {
   runApp(App());
