@@ -5,9 +5,8 @@ class ImageModel {
 
   ImageModel(this.id, this.url, this.title);
 
-  ImageModel.fromJson(Map<String, dynamic> parsedJson) {
-    id = parsedJson['id'];
-    url = parsedJson['url'];
-    title = parsedJson['title'];
-  }
+  ImageModel.fromJson(Map<String, dynamic> parsedJson)
+      : id = parsedJson['id'],
+        url = parsedJson['url'],
+        title = parsedJson['title'];
 }
